Add files via upload
This is a neural network built from scratch without using any libraries. This is my first experience with neural networks and I've done it by following a tutorial which helped me understand and learn about neural networks in depth. This will be the first of many projects to be added soon.